views: Move one-click unsubscribe to its own file.
